+++ pyNastran/op2/op2_interface/op2_reader.py
@@ -4063,9 +4063,15 @@ class OP2Reader:
             raise NotImplementedError(f'data={data!r}; n={len(data)}')
         return table_name.strip()
 
-    def _read_table_name(self, rewind=False, stop_on_failure=True) -> bytes:
+    def _read_table_name(self, last_table_name: Optional[bytes]=None,
+                         rewind: bool=False, stop_on_failure: bool=True) -> bytes:
         """
         Reads the next OP2 table name (e.g. OUG1, OES1X1)
+
+        Parameters
+        ----------
+        last_table_name : bytes; default=Noen
+            the last table name
 
         Returns
         -------
@@ -4111,6 +4117,8 @@ class OP2Reader:
                     if not is_special_nastran and op2.post != -4:
                         op2.f.seek(op2.n)
                         self.show(1000)
+                        if last_table_name:
+                            self.log.error(f'finished table_name = {last_table_name}')
                         raise FatalError('There was a Nastran FATAL Error.  Check the F06.\n'
                                          f'last table={op2.table_name!r}; post={op2.post} '
                                          f'version={self.op2._nastran_format!r}')
@@ -4975,7 +4983,8 @@ class OP2Reader:
         if hasattr(self, 'obj'):
             str(self.obj.get_stats())
 
-    def show(self, n: int, types: str='ifs', endian: Optional[str]=None):  # pragma: no cover
+    def show(self, n: int, types: str='ifs', endian: Optional[str]=None,
+             force: bool=False):  # pragma: no cover
         """
         shows the next N bytes
 
@@ -4987,16 +4996,19 @@ class OP2Reader:
             the data types to show
         endian : str; default=None -> active endian
             the data endian
+        force : bool; default=False
+            overwrite the n=2000 limiter
 
         """
         op2 = self.op2
         assert op2.n == op2.f.tell()
         data = op2.f.read(n)
-        strings, ints, floats = self.show_data(data, types=types, endian=endian)
+        strings, ints, floats = self.show_data(data, types=types, endian=endian, force=force)
         op2.f.seek(op2.n)
         return strings, ints, floats
 
-    def show_data(self, data: bytes, types: str='ifs', endian: Optional[str]=None):  # pragma: no cover
+    def show_data(self, data: bytes, types: str='ifs', endian: Optional[str]=None,
+                  force: bool=False):  # pragma: no cover
         """
         Shows a data block as various types
 
@@ -5017,14 +5029,17 @@ class OP2Reader:
             Q - unsigned long long (int; 8 bytes)
         endian : str; default=None -> auto determined somewhere else in the code
             the big/little endian {>, <}
+        force : bool; default=False
+            overwrite the n=2000 limiter
 
         .. warning:: 's' is apparently not Python 3 friendly
 
         """
         #ifsdqlILQ
-        return self._write_data(sys.stdout, data, types=types, endian=endian)
+        return self._write_data(sys.stdout, data, types=types, endian=endian, force=force)
 
-    def _write_data(self, f, data: bytes, types: str='ifs', endian: Optional[str]=None):  # pragma: no cover
+    def _write_data(self, f, data: bytes, types: str='ifs',
+                    endian: Optional[str]=None, force: bool=False):  # pragma: no cover
         """
         Useful function for seeing what's going on locally when debugging.
 
@@ -5045,10 +5060,13 @@ class OP2Reader:
             Q - unsigned long long (int; 8 bytes)
         endian : str; default=None -> auto determined somewhere else in the code
             the big/little endian {>, <}
+        force : bool; default=False
+            overwrite the n=2000 limiter
 
         """
         n = len(data)
-        if n > 2000:
+        if not force and n > 2000:
+            self.log.warning(f'limiting n={n} to 2000')
             n = 2000
         nints = n // 4
         ndoubles = n // 8