+++ pyNastran/op2/writer/op2_writer.py
@@ -15,6 +15,10 @@ from .geom3_writer import write_geom3
 from .geom4_writer import write_geom4
 from .ept_writer import write_ept
 from .mpt_writer import write_mpt
+#from .edt_writer import write_edt
+#from .edom_writer import write_edom
+#from .dit_writer import write_dit
+#from .dynamic_writer import write_dynamic
 if TYPE_CHECKING:  # pragma: no cover
     from pyNastran.op2.op2 import OP2
 
@@ -97,20 +101,25 @@ def _write_op2(fop2, fop2_ascii, obj: OP2,
     struct_3i = Struct(endian + b'3i')
     write_op2_header(obj, fop2, fop2_ascii, struct_3i, post=post, endian=endian)
 
-    if 'GEOM1' not in skips:
+    if 'GEOM1' not in skips:  # nodes
         write_geom1(fop2, fop2_ascii, obj, endian=endian)
-    if 'GEOM2' not in skips:
+    if 'GEOM2' not in skips:  # elements
         write_geom2(fop2, fop2_ascii, obj, endian=endian)
-    if 'GEOM3' not in skips:
+    if 'GEOM3' not in skips:  # constraints
         write_geom3(fop2, fop2_ascii, obj, endian=endian, nastran_format=nastran_format)
-    if 'GEOM4' not in skips:
+    if 'GEOM4' not in skips:  # loads
         write_geom4(fop2, fop2_ascii, obj, endian=endian, nastran_format=nastran_format)
-    if 'EPT' not in skips:
+    if 'EPT' not in skips:    # properties
         write_ept(fop2, fop2_ascii, obj, endian=endian)
-    if 'MPT' not in skips:
+    if 'MPT' not in skips:    # materials
         write_mpt(fop2, fop2_ascii, obj, endian=endian)
-    #if 'DIT' not in skips:
-        #write_dit(fop2, fop2_ascii, obj)
+
+    #if 'EDT' not in skips:  # aero
+        #write_edt(fop2, fop2_ascii, obj, endian=endian)
+    #if 'EDOM' not in skips:  # optimization
+        #write_edom(fop2, fop2_ascii, obj, endian=endian)
+    #if 'DIT' not in skips:  # tables
+        #write_dit(fop2, fop2_ascii, obj, endian=endian)
     #if 'DYNAMIC' not in skips:
         #write_dynamic(fop2, fop2_ascii, obj)
     if 'grid_point_weight' not in skips:
@@ -272,7 +281,7 @@ def _write_result_tables(obj: OP2, fop2, fop2_ascii, struct_3i, endian, skips: S
                     continue
             else:
                 raise NotImplementedError(f'  *op2 - {result.__class__.__name__} not written')
-                #obj.log.warning("  *op2 - %s not written" % result.__class__.__name__)
+                obj.log.warning(f'  *op2 - {result.__class__.__name__} not written')
                 #continue
 
             case_count += 1